(feat)(fix) Created a POST request in order to recieve the Availability
This commit has modified the following files;
	Requests.cs
	RestaurantPage.xaml
	RestaurantPage.xaml.cs
The following file was created;
	AvailableTimes.cs
In the Requests.cs file a new method was added, this method when called with a url, token, and parameters of the body of the message will perform a POST call to the API and return the available times for the next week - subject to change.
Bug Fixes were made in regards to the app crashing with no WiFi.

Unfortunately a new bug has arisen that means the POST request does not occasionaly return the correct values - this will be corrected in future updates.