(feat) Display Current Promotions for Time Slot
The application will now display, the current available promotions for each timeslot. This will be displayed on the card viewer alongside the date and time of the slot. AvailableTimes.cs, RestaurantPage.xaml, and RestaurantPAge.xaml.cs.

This has caused a bug where when there is no time slot available the card viewer will not display - this will be fixed in a future iteration