feat(UI): UI re-design
the UI has been re-designed so that the availability has been added to a different tab making it look a lot cleaner. scan again button has been removed as it doesn't make sense to have it in the tab bar. the home screen will now show about and contact details instead of availability