(feat) Added a feature to change the Party Size of the User
Implemented a slider within the booking tab which the user can interact with to change the booking size for there party and the app will find all the available tables for a booking of that size. There are still some problems with the formatting of the page - requires changes to the xaml code.