(feat)(fix) Added a slider for the user to change the party size
Changes were made to the Booking tab of the Main page to display a slider at the bottom of the booking page for the user to adjust to change the party size, default value = 3. The page will then create a new call with the changed party size and display the updated values. If no restaurants are available an error message will be displayed.