feat(reviews page): added no reviews label
added a label which displays when there are no reviews which looks much nicer than just having an empty list view. made some alterations to the reviews pop too so that it is more informative. also added a view ratings label which signifies that the review can be interacted with

close #115