(fix) Updated Icons - Fixed Camera Page Crashing Bug
Updated the icons displayed on both iOS and Android. They now are of higher resolution so look more appealing to the user.
Also fixed the bug discovered by Scott. This was due to the apps camera Page not being able to handle the back trace of the phone. this was fixed by preventing new pages to be added to the back trace, a try catch statement was also used so the camera restarts if an exception is still thrown.